requirements.txt
streamlit>=1.37.0
openai>=1.40.0
python-dotenv>=1.0.1
pymupdf>=1.24.9
chromadb>=0.5.5
sentence-transformers>=3.0.1
pandas>=2.2.2
numpy>=1.26.4
reportlab>=4.2.2
pydantic>=2.8.2
tqdm>=4.66.4
